==> prior/05b_convert_extracted_math.ipynb <==
%load_ext autoreload
%autoreload 2

import csv
import ipywidgets as widgets
import json
import openai
import os
import pandas as pd
import papermill as pm
import scrapbook as sb

from IPython.display import display

from functions import gpt

from tqdm.notebook import tqdm_notebook
tqdm_notebook().pandas()

pd.set_option('display.max_colwidth', None)

base_name = "2021_Hashimoto_Neural_ODE_and_holographic_QCD_PUB"
project_folder = "diygenomics-projects"
sub_category = "math"
work_bucket = "AdS-CFT"
external_id = "2023_05_22_92dc0613b4493d7b5847g"

data_path = os.getenv('DATA_PATH')
file_path = lambda *args: os.path.join(data_path, project_folder, sub_category, work_bucket,
                                       base_name, 'mathpix', *args)

model = 'gpt-4' # 'gpt-3.5-turbo' # 'gpt-4'
index_col = 'uuid'

openai_model = 'text-embedding-ada-002'
# openai_embeddings = OpenAIEmbeddings()

system_prompt = """You are an academic researcher and mathmatician in information systems biology.
- Convert the Latex expression into python using sympy: 
- Do not output anything except code.
- Do not provide an explaination.
- Do not include comments.
"""

df = pd.read_csv(file_path('extracted_math.csv'), index_col=index_col)

df_display = df[df['type'] == 'display']

general_errors = {}

def generate_python_math(row):  
    response = None

    if len(row['math']) > 0:
        possible_response = gpt.chat_create(system_prompt, row['math'], model, output_json=False)
        if 'general_error' in possible_response:
            general_errors[row.name] = response
        else:
            response = possible_response
        
    return response

df_display['gpt_python_math'] = df_display.progress_apply(lambda row: generate_python_math(row), axis=1)

df_display['openai_math_embeddings'] = df_display['math'].progress_apply(lambda x: openai.Embedding.create(input=x, engine=openai_model, temperature=0,)['data'][0]['embedding'])

df_display['openai_sympy_embeddings'] = df_display['gpt_python_math'].progress_apply(lambda x: openai.Embedding.create(input=x, engine=openai_model, temperature=0,)['data'][0]['embedding'])

df_display[['math']]

df_display.to_csv(file_path('extracted_display_math.csv'), quoting=csv.QUOTE_MINIMAL)

# convert this to python using sympy:

# \mathrm{d} s^{2}=-f(\eta) \mathrm{d} t^{2}+\mathrm{d} \eta^{2}+g(\eta)\left(\mathrm{d} x_{1}^{2}+\cdots+\mathrm{d} x_{d-1}^{2}\right)

# do not output anything except code. do not provide an explaination.
